UPDATE REWORDING FILENAMES:
Rewording filenames to be more explicit in parsing directory.
