+++ srcs/parsing/check_ceilling_floor_texture_functions.c
