+++ srcs/parsing/check_map_content.c
