+++ srcs/parsing/check_wall_texture_functions.c
